+++ av1an/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+import vmaf
+import argparse